Fixes Suggestions

+++ src/main/java/ca/sheridancollege/repositories/RecipeRepository.java
@@ -32,7 +32,7 @@ public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 
 	@Query
 	(value = "CALL SuggestRecipe(:user_id)" , nativeQuery = true)
-	public List<Recipe> suggestRecipes(@Param("user_id") int user_id);
+	public List<Recipe> suggestRecipes(@Param("user_id") long user_id);
 	
 	@Query
 	(value = "CALL BasicSearch(:searchInput)" , nativeQuery = true)
@@ -40,18 +40,18 @@ public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 	
 	@Query
 	(value = "CALL SuggestDiet(:user_id)" , nativeQuery = true)
-	public List<Recipe> suggestDiet(@Param("user_id") int user_id);
+	public List<Recipe> suggestDiet(@Param("user_id") long user_id);
 	
 	@Query
 	(value = "CALL SuggestCuisine(:user_id)" , nativeQuery = true)
-	public List<Recipe> suggestCuisine(@Param("user_id") int user_id);
+	public List<Recipe> suggestCuisine(@Param("user_id") long user_id);
 	
 	@Query
 	(value = "CALL SuggestCountry(:user_id)" , nativeQuery = true)
-	public List<Recipe> suggestCountry(@Param("user_id") int user_id);
+	public List<Recipe> suggestCountry(@Param("user_id") long user_id);
 	
 	@Query
 	(value = "CALL SuggestProtein(:user_id)" , nativeQuery = true)
-	public List<Recipe> suggestProtein(@Param("user_id") int user_id);
+	public List<Recipe> suggestProtein(@Param("user_id") long user_id);
 
 }
